requirements.txt
cryptography
python-dotenv
secrets
pyperclip